Implement Prolific monitor.
Prolific is going to be a meaningful alternative because it is going to
work hard to capture the final messages from your program that are lost
in naive attempts at log shuttling. It will also minimize its footprint
in your application, no logging network communication within your
application, that is all deferred.