Set logging fd via environment.
Use `PROLIFIC_LOGGING_FD` to pass the watched process the file
descriptor number that is used to gather logging information. The
watched process will configure a log shuttle that writes logs to the
specified file descriptor.

This is a good use of environment. It is used this way within Node.js.
It is a way of communicating information about the environment; the
logging environment. I work hard to discourage the abuse of the
environment to pass parameters, but this raw bootstrapping information
that has more to do with an environmental system than with the program
itself is what environment variables are for. It is analogous to `HOME`,
`USER`, `TERM` or `DISPLAY`.