Exit callback is necessary.
Necessary for uncaught exception when the default shutdown hooks in
Node.js are disabled by a `SIGTERM` handler. My test to assert that we'd
shutdown normally passed because the only hook was the async pipe
between the parent and child. That is closed by `exit`. Anything else is
going to keep the socket open, so you need to exit explicitly. If you
exit immediately after write, then you will not flush STDERR.

Put this note somewhere where you'll not delete it.