Create a simple test queue.
The queue will gather log entries into an array and proves a means to
seek specific entries returning them for comparison. Not sure how to
deal with the date stamp. Might add the ability to override specific
values, to merge or overlay an object when shifting.

The `prolific.sink` does expose its `Date` object for replacement for
testing, so I'll probably stick with that. We're already setting
properties at the outset. The date stamp should be the only thing that
Prolific generates that is non-deterministic.