Operate on array.
A premature optimziation; I don't want to build the flattened log entry
if it is not going to match and it is probably not going to match on the
level mostly, so let's wait to see if the tests pass before we flatten.

We'll flatten on the listener process side.