Seek Unicode character, notes on buffer lengths.
Resuming the use of a final message on standard error to close down the
monitor. Not sure why I restored to using IPC to communicate the
shutdown but it was probably because either a) I was disheartened by the
sight of chunk messages getting interleaved with stack trace spew in
standard error that the Prolific Collector could not parse and b) I was
using IPC anyway to transmit updated logging pipelines.

Now I've found that fatal stack traces are not logged. The monitor has
shut down ordering problems. It really should be shut down from within
the pipeline processor when the processor detects a shutdown message.
This message should be akin to the pipeline version message. We're
already sending meta messages to indicate a change in contents of the
stream, messages that are pre-triage-update and messages that are
post-triage-update.

To make this change I've considered preserving the process tree modeling
in order to ensure that all child processes are accounted for, but we'll
know they're accounted for when all of the monitors shut down.

Ultimately the monitors, if they are shared by processes, are going to
have to keep a process count. Something to consider when I do multiple
monitors. We're reducing the complexity in the root process and
increasing it in the monitor.