Prepend newline to chunk.
By prepending a newline we ensure that a chunk will always start on a
newline. This means that if our chunk is interleaved with the output of
another process on standard error that we'll be able to extract our
chunk from standard out and faithfully reproduce the output of the other
process.

While we might be carefully emitting chunks that fit within `PIPE_BUF`,
other processes might not be so careful. Output to standard error, like
long or elaborate stack traces, could be jagged and broken, but we'll
still successfully extract our chunks and leave standard error output no
more jagged and broken than any other multi-process Node.js application.

This bit of StackOverflow talks about where we're headed.

https://serverfault.com/a/733611