Simplify chunk.
Chunk is now entirely fixed width. Chunks will only be used on the
standard error stream. Not going to tell myself that using them on in
the pipe is good exercise. Reduces complexity. Might get rid of
Prolific Consolidator entirely since we only need to use `byline` to
read the socket in the monitor.

Chunks are now either control chunks or body content chunks. In fact,
maybe I should rename `header` to `control` so that I don't use the word
header quite so much.

Updated description of the header in the comments of `chunk.js`.