Regex by line based collector.
Rather than scanning and scanning looking for `%` we split the stream by
lines and apply a regular expression to the line. Greatly simplifies the
scanner logic.

We are no longer using chunks for the asynchronous stream so no more
tests to determine if there is extra stuff in the stream when the stream
is our dedicated pipe. Used to be concerned that the forgiveness we
afford standard error would be given to our dedicated pipe and it would
forgive the unforgivable.

Still synchronous but easy to wrap in a loop that reads from standard
error.