Change example entries.
Updated the example of entries so that when I see them I'm not wondering
why it is an array of array and not an array of a single object. It
looks more like what entries are supposed to look like and doesn't make
me wonder if Prolific Queue is formatting entries incorrectly.