Rename `chunk` event to `entries`.
Used to be that we where using chunks to talk to the monitor through the
dedicated asynchronous pipe as well as over standard out so that when we
where sending entries we where actually sending chunks. Chunks where
within chunks. No longer. Thus, let's have a better name for that event
and not confuse ourselves when we revisit this code in the future.