Simplified consolidator.
There is still a place for the Consolidator in Prolific. We need to know
whether we're repeating messages that have been sent by both the
asynchronous pipe and the synchronous standard error tunneling. We can
see that the synchronous final messages might overlap with messages that
where successfully received via the asynchronous pipe and avoid
duplicating them.