Restore final message capture.
There are many notes in the most recent commits about this problem. I
was terminating everything in the monitor at once, including the
processing pipeline when the fatal messages arrived.

There where other timing issues as well, but they've all been resolved
or are capable of being resolved.

Greatly simplified inter-process communication. Only using the
checksummed chunks to tunnel messages through standard error. Notes on
progress on this work is in the commit comments for Prolific Chunk,
Prolific Queue, and Prolific Collector.

Removed the `closer.js` and no longer modeling the process tree. The
child must write its exit out to standard out which it will do if it has
been configured correctly, even when it fails. Only if it gets a
`SIGKILL` or the like, but if you're going to `SIGKILL` a child process
please just kill the whole program.

Monitor will now shutdown an exit message passes through the processor
which will only happen when it gets one from its standard input or when
standard input closes. The supervisor will shutdown monitors by closing
their standard input instead of sending `SIGTERM`.