Upgrade Arguable to 8.0.x.
Upgrade Arguable to 8.0.x. It is now Destructor based and provides the
destructor that we always create each time we use Arguable.

It also provides a way to test pipes without having to create a file
descriptor attached to a socket. Simplifies testing the asynchronous
logging pipe without having to create mocks.