Create a `stackTrace` reporter.
Need to be able to log errors without having to generate errors in unit
tests, especially for `events.EventEmitter` error events. Created a
`stackTrace` function builder that will create a function that accepts
an error and logs it. Specifying the level is optional. Additional
context is also optional.

The logger will create a logging message with `Error.stack` and
`Error.code` properties. For Node.js this captures the results of most
system errors that I'm aware of. I use the Interrupt library a lot and
it prints its context into its stack trace, so I don't want to add all
of the error properties since they would be duplicated. You can chance
this behavior with an optional properties array.