Process arrays of entries, assemble in processor.
Assemble in the processor instead of assembling along the way. The
processor now gets an array of entry objects with the component parts of
an entry as properties to assemble how it sees fit.

Invocation of triage and processor is now consistent pre- and post-
triage injection as well as for processor local messages.

Gather entries from the stream into an array. Not going to send gathered
arrays from the shuttle side at this point.

Closes #999.
Closes #998.