Build processor in `try`/`catch`.
Building the processor in `try`/`catch` block. Reconfiguration is
occasionally failing in test. Assume that is because it cannot
completely read the file. Hope that this means that there will be a
subsequent event where it can read the file. Otherwise, maybe it makes
sense to copy the processor to the temp directory, or maybe you retry in
a few seconds on failure to see if the rest of the file has been
written. Those two would be the same thing.

Actually, you probably want to checksum the file a couple times until
you determine that the checksums are stable, if you can get two
checksums in a row, the file is ready to read, it has stabilized. You'd
want to give it enough time to flush to disk, not read it immediately in
a loop. Maybe you need to get a file descriptor and tail the file.

Check that you're able to delete a file and still read it from a file
descriptor on the operating systems you care about.

https://support.hpe.com/hpsc/doc/public/display?docId=emr_na-c00833030

See #1027.