Fix unit test on Linux.
Shuttle test failed because `Shuttle` sets the logging sink and
`Watcher` was logging that a check sum failed (a file had been deleted.)
Not sure why `Watcher` was even attempting to read the file, but I fixed
the test by returning a slice of the gathered messages, instead of
turning the gathered messages.

See #1037.