`Watcher` checksum bugs, log through `EventEmitter`.
`Watcher` now logs through `EventEmitter`, not the Prolific Logger.
Can't use it to test Prolific Shuttle because Prolific Shuttle sets the
Prolific Sink and the Prolific Watcher starts to log itself through
itself.

The regular expression that extracts the checksum had no capture. It was
a freak of number parsing that allowed it to pass unit tests when the
hash value was always zero.

`Watcher` would attempt to parse JSON even after the checksum failed.

Closes #1041.
Closes #1040.
Closes #1039.