`uref` IPC channel, for now.
The IPC channel prevents the event loop from ending. We're trying to
have Prolific communicate with the sidecar using modes of communication
that can be `uref` or otherwise to not prevent the event loop from
exiting. Node.js IPC is global. This is not an acceptable invasion of
the host application environment.

Closes #1049.