`Watcher` emits end-of-stream on killed.
The `Killer` in Prolific Executable will detect the application process
exit, but we need to make sure we've consumed all the synchronous writes
in the temporary directory before we actually shutdown the sidecar. The
`Watcher` now has a `killed()` method that takes a process id and it
will then emit an end-of-stream meesage when the temporary directory has
no more synchronous write files for that process id.

Closes #1059.