Layered shutdown for logging printer.
We now have a logging printer so we're shutting down in layers. We don't
want to exit the moment we know that all of our child processes have
exited. We want to flush our logs to the temporary directory, then wait
until the temporary directory is empty. The supervisor could decide to
log directly to the printer at some point, which might make shutdown
slightly less complicated.

See #1121.