Working logging of Prolific itself.
On the right track. It occurs to me that we could also embed a
processor in the supervisor and have it be the one to log all the meta
messages through the same processor definition. Bulks up the monitor,
a bit though, doesn't it? Wonder if I'm ever going to want to come back
and put Prolific on a diet.

Also, probably going to move our Logger into Prolific Queue so we can
have the same trace messaging in Prolific Queue.

Considering allowing a logger to accept a custom `sink` so that we can
just reuse the Prolific Logger to triage, format and stamp messages.

See #1121.