Socket destroyed while waiting in queue.
Socket arriving null in sidecar because it is destroyed while waiting
for the sidecar to start. Numerous race conditions here. Olio is where I
first applied socket passing. We don't have these problems there because
we generally expect an Olio program to start and if it doesn't start the
success or failure of socket passing isn't that important. Here we do
expect to have failed connections, failing at any stage.

It seems like it is benign to pass a closed socket. It gets converted to
`null` successful. If `Descendent` has shutdown then `Descendent` should
probably just drop the message. We can't expect it to be delivered
through a process tree, it's not a guaranteed delivery channel, it is
fire-and-forget.

See #1121.