Do not rename processor if it is in search path.
If the user specifies an absolute path, we assume that it is not in the
search path, and anyway we won't be able to search for it, so we do
rename it to place it next to the executable. From there it ought to be
able to find its necessary resources. We need to tell the user to keep
the processors next to the main method. We can throw in some switches to
either change the "mount" path or disable this behavior.

Closes #1147.