Construct Logger class with Sink parameter.
Make the Prolific Sink a parameter to Prolific Logger. Now we can reuse
the logger to do meta-logging of Prolific itself by creating a separate
logger within the supervisor and sidecar. In the sidecar especially we
need to allow the user to log messages from the processor, while we'll
be wanting to log messages about the state of sidecar.

See #1165.