Polling pids only to detect exit.
Only polling pids to detect exit. Removed the exit message from Prolific
Queue and broke the shutdown logic. Here is a sketch of pid polling that
gets it working again. Might be better Happenstance driven, but I don't
want to do a rewrite at the moment, nor add new dependencies.

We're now counting on being able to detect the exit with the socket
closure. The user will no longer be closing the socket themselves, and
we're not expecting a UNIX domain socket to hang like a TCP socket. We
might be mistaken about the hanging.

Closes #1213.